\name{biSBM-package}
\alias{biSBM-package}
\alias{biSBM}
\docType{package}
\title{
\packageTitle{biSBM}
}
\description{
\packageDescription{biSBM}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{biSBM}
\packageIndices{biSBM}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{biSBM}

Maintainer: \packageMaintainer{biSBM}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
